moved old geometry engine to the attic
